requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
idna==3.10
joblib==1.4.2
kafka-python==2.0.2
nltk==3.9.1
oauthlib==3.2.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
textblob==0.18.0.post0
tqdm==4.66.5
tweepy==4.14.0
urllib3==2.2.3
